+++ resources/js/app.js
@@ -4,9 +4,9 @@
  * building robust, powerful web applications using Vue and Laravel.
  */
 
-require('./bootstrap');
+//require('./bootstrap');
 
-window.Vue = require('vue').default;
+//window.Vue = require('vue').default;
 
 /**
  * The following block of code may be used to automatically register your
@@ -19,7 +19,7 @@ window.Vue = require('vue').default;
 // const files = require.context('./', true, /\.vue$/i)
 // files.keys().map(key => Vue.component(key.split('/').pop().split('.')[0], files(key).default))
 
-Vue.component('example-component', require('./components/ExampleComponent.vue').default);
+//Vue.component('example-component', require('./components/ExampleComponent.vue').default);
 
 /**
  * Next, we will create a fresh Vue application instance and attach it to
@@ -27,7 +27,32 @@ Vue.component('example-component', require('./components/ExampleComponent.vue').
  * or customize the JavaScript scaffolding to fit your unique needs.
  */
 
-const app = new Vue({
-    el: '#app',
-});
-/* @todo: Rework */
+//const app = new Vue({
+//    el: '#app',
+//});
+
+/* @todo: Rework Register automatically */
+
+/**
+ * First we will load all of this project's JavaScript dependencies which
+ * includes Vue and other libraries. It is a great starting point when
+ * building robust, powerful web applications using Vue and Laravel.
+ */
+
+require('./bootstrap');
+
+// Load vue 
+window.Vue = require('vue');
+
+// Create the app (without any parameters }
+const app = Vue.createApp({});
+
+// Import the component
+import ExampleComponent from './components/ExampleComponent.vue'
+
+// Register the component
+app.component('example-component', ExampleComponent);
+
+// Mount the app 
+app.mount('#app');
+