+++ CHANGELOG.md
@@ -3,12 +3,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## unreleased
+## ## 2021-01-10
+
 
 ### Added
 
+- Upgrade Vue to Vue 3.03 (was not 3.03 yet...)
+- Enable Vue with the one example component (=> show it in home)
+- Adapt gitignore 
 
-## 2021-01-10
+## 2021-01-09
 
 ### Added